Fix Windows build after 25373
R=bmeurer@chromium.org
TBR=bmeurer@chromium.org

Review URL: https://codereview.chromium.org/733783003

Cr-Commit-Position: refs/heads/master@{#25374}

+++ src/compiler/node-matchers.h
@@ -190,7 +190,7 @@ struct AddMatcher : public BinopMatcher {
         typename BinopMatcher::RightMatcher::ValueType value =
             m.right().Value();
         if (value >= 0 && value <= 3) {
-          return value;
+          return static_cast<int>(value);
         }
       }
     } else if (node->opcode() == kMulOpcode) {
